requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
PyJWT==2.9.0
supabase==2.9.0
google-generativeai==0.8.0
stripe==10.0.0
httpx==0.27.0
python-multipart==0.0.9
slowapi==0.1.9
structlog==24.4.0
celery[redis]==5.4.0
redis==5.1.0
sentry-sdk[fastapi]==2.14.0
pytest==8.3.0
pytest-asyncio==0.24.0
